[dy] Address comments
